requirements.txt
flask==3.1.0
flask-cors==5.0.0
anthropic==0.49.0
python-dotenv==1.0.1
gunicorn==21.2.0
